readme: Remove.
One can always read the readme at `zalmoxisus/remotedev-serialize`,
and this makes it faster to develop our fork. In particular, we
won't have to keep the `Supported` list up to date as we change the
list of Immutable data structures that our fork handles. (We'll go
straight to the code for that; if we get something wrong, a stack
trace will be more helpful than a README.)